+++ plugins/arrow-flight-rpc/src/main/java/org/opensearch/arrow/flight/transport/FlightStreamPlugin.java
@@ -364,7 +364,8 @@ public class FlightStreamPlugin extends Plugin
                 ServerComponents.SETTING_FLIGHT_PORTS,
                 ServerComponents.SETTING_FLIGHT_HOST,
                 ServerComponents.SETTING_FLIGHT_BIND_HOST,
-                ServerComponents.SETTING_FLIGHT_PUBLISH_HOST
+                ServerComponents.SETTING_FLIGHT_PUBLISH_HOST,
+                ServerComponents.SETTING_FLIGHT_PUBLISH_PORT
             )
         ) {
             {