+++ test/framework/src/main/java/org/opensearch/search/aggregations/AggregatorTestCase.java
@@ -501,6 +501,7 @@ public abstract class AggregatorTestCase extends OpenSearchTestCase {
         when(searchContext.bitsetFilterCache()).thenReturn(new BitsetFilterCache(indexSettings, mock(Listener.class)));
         IndexShard indexShard = mock(IndexShard.class);
         when(indexShard.shardId()).thenReturn(new ShardId("test", "test", 0));
+        when(indexShard.indexSettings()).thenReturn(indexSettings);
         when(searchContext.indexShard()).thenReturn(indexShard);
         SearchOperationListener searchOperationListener = new SearchOperationListener() {
         };